Purge IPFS code, reorg netsim cmd
