+++ cmd/netsim/main.go
@@ -50,6 +50,8 @@ func getRandomInt(m int) int {
 	return r
 }
 
+
+
 func main() {
 	if len(os.Args) < 2 {
 		log.Crit("please specify number of nodes to start in simulation: go run main.go [num]")
@@ -61,17 +63,18 @@ func main() {
 		log.Crit("main", "error", err)
 	}
 
-	sim, err := p2p.NewSimulator(num)
+	sim, err := NewSimulator(num)
 	if err != nil {
 		log.Crit("NewSimulator", "error", err)
 	}
 
-	defer func() {
-		err = sim.IpfsNode.Close()
-		if err != nil {
-			log.Warn("main", "warn", err.Error())
-		}
-	}()
+	// TODO: Since we removed IPFS this needs to handle error channel from sim.Bootnode.Stop()
+	//defer func() {
+	//	err = sim.IPFSNode.Close()
+	//	if err != nil {
+	//		log.Warn("main", "warn", err.Error())
+	//	}
+	//}()
 
 	for _, node := range sim.Nodes {
 		e := node.Start()
