+++ p2p/address_utils.go
@@ -1,0 +1,28 @@
+package p2p
+
+import (
+	"github.com/libp2p/go-libp2p-core/peer"
+	ma "github.com/multiformats/go-multiaddr"
+
+)
+
+func stringToPeerInfo(peerString string) (peer.AddrInfo, error) {
+	maddr, err := ma.NewMultiaddr(peerString)
+	if err != nil {
+		return peer.AddrInfo{}, err
+	}
+	p, err := peer.AddrInfoFromP2pAddr(maddr)
+	return *p, err
+}
+
+func stringsToPeerInfos(peers []string) ([]peer.AddrInfo, error) {
+	pinfos := make([]peer.AddrInfo, len(peers))
+	for i, p := range peers {
+		p, err := stringToPeerInfo(p)
+		if err != nil {
+			return nil, err
+		}
+		pinfos[i] = p
+	}
+	return pinfos, nil
+}
