+++ p2p/p2p_test.go
@@ -69,37 +69,6 @@ func TestGenerateKey(t *testing.T) {
 	}
 }
 
-func TestStart(t *testing.T) {
-	ipfsNode, err := StartIpfsNode()
-	if err != nil {
-		t.Fatalf("Could not start IPFS node: %s", err)
-	}
-
-	defer ipfsNode.Close()
-
-	ipfsAddr := fmt.Sprintf("/ip4/127.0.0.1/tcp/4001/ipfs/%s", ipfsNode.Identity.String())
-
-	t.Log("ipfsAddr:", ipfsAddr)
-
-	testServiceConfig := &Config{
-		BootstrapNodes: []string{
-			ipfsAddr,
-		},
-		Port: 7001,
-	}
-
-	s, err := NewService(testServiceConfig)
-	if err != nil {
-		t.Fatalf("NewService error: %s", err)
-	}
-
-	e := s.Start()
-	err = <-e
-	if err != nil {
-		t.Errorf("Start error: %s", err)
-	}
-}
-
 func TestService_PeerCount(t *testing.T) {
 	testServiceConfigA := &Config{
 		NoBootstrap: true,
@@ -215,25 +184,5 @@ func TestSend(t *testing.T) {
 	err = sa.Send(p, msg)
 	if err != nil {
 		t.Errorf("Send error: %s", err)
-	}
-}
-
-func TestNoBootstrap(t *testing.T) {
-	testServiceConfigA := &Config{
-		NoBootstrap: true,
-		Port:        7006,
-	}
-
-	sa, err := NewService(testServiceConfigA)
-	if err != nil {
-		t.Fatalf("NewService error: %s", err)
-	}
-
-	defer sa.Stop()
-
-	e := sa.Start()
-	err = <-e
-	if err != nil {
-		t.Errorf("Start error: %s", err)
 	}
 }
